stub out audio monster

+++ test/processors/audio_processor_test.rb
@@ -2,6 +2,10 @@ require 'test_helper'
 
 class AudioProcessorTest < ActiveSupport::TestCase
 
+  def travis?
+    ENV['TRAVIS']
+  end
+
   before {
     WebMock.disable_net_connect!
   }
@@ -10,7 +14,15 @@ class AudioProcessorTest < ActiveSupport::TestCase
     WebMock.allow_net_connect!
   }
 
-  let(:processor) { AudioProcessor.new(logger: Logger.new('/dev/null')) }
+  let(:audio_monster) do
+    Minitest::Mock.new
+  end
+
+  let(:processor) {
+    p = AudioProcessor.new(logger: Logger.new('/dev/null'))
+    p.audio_monster = audio_monster if travis?
+    p
+  }
 
   it 'defines supported tasks' do
     AudioProcessor.supported_tasks.first.must_equal 'transcode'
@@ -35,6 +47,12 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it 'should return analysis with loudness' do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:info_for_wav, {:size=>4277159, :content_type=>"audio/vnd.wave", :channel_mode=>"Mono", :bit_rate=>705, :length=>48, :sample_rate=>44100}, [String])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+      end
+
       processor.on_message(msg)
       processor.result_details[:info][:length].to_i.must_equal 48
       processor.result_details[:info][:loudness][:integrated_loudness][:i].must_equal -18.5
@@ -56,6 +74,14 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it 'should return sliced duration' do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        # audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+        audio_monster.expect(:slice_wav, "/tmp/audio_monster/test_long.wav20150528-43657-18y5zb.wav", [String, String, String, String])
+        audio_monster.expect(:info_for_wav, {:size=>441044, :content_type=>"audio/vnd.wave", :channel_mode=>"Mono", :bit_rate=>705, :length=>5, :sample_rate=>44100}, [String])
+      end
+
       processor.on_message(msg)
       processor.result_details[:info][:length].to_i.must_equal 5
     end
@@ -76,6 +102,13 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it 'should return one tone at 5-6 seconds' do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        # audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:encode_wav_pcm_from_flac, ["0\n", ""], [String, String])
+        audio_monster.expect(:tone_detect, [{:start=>5.055, :finish=>6.2, :min=>0.050196286, :max=>0.091354366}], [String, Fixnum, Float, Float])
+      end
+
       processor.on_message(msg)
 
       # puts 'tones:'
@@ -104,6 +137,13 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it "should return one silence period, from 6-15" do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        # audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:encode_wav_pcm_from_flac, ["0\n", ""], [String, String])
+        audio_monster.expect(:silence_detect, [{:start=>6.168, :finish=>14.999, :min=>0.0, :max=>0.0049678111}], [String, Float, Float])
+      end
+
       processor.on_message(msg)
 
       # processor.result_details[:info].each{|t| puts "#{t[:start].to_i.to_time_string_summary} - #{t[:finish].to_i.to_time_string_summary}" }
@@ -144,26 +184,36 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it "should set destination format" do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        # audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:create_wav_wrapped_mp2, true, [String, String, Hash])
+        audio_monster.expect(:info_for_wav, {:size=>4277159, :content_type=>"audio/vnd.wave", :channel_mode=>"Mono", :bit_rate=>705, :length=>48, :sample_rate=>44100}, [String])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+      end
+
       processor.on_message(msg)
       processor.destination.wont_be_nil
       processor.destination_format.must_equal 'wav'
     end
 
     it "should create a wave wrapped mp2" do
-      result_file_path = out_file('fixer/test.fixer.org/test_short_wrap.wav')
-      FileUtils.rm_f(result_file_path)
-      processor.on_message(msg)
-      File.exists?(result_file_path).must_equal true
-      wave = NuWav::WaveFile.parse(result_file_path)
+      if !travis?
+        result_file_path = out_file('fixer/test.fixer.org/test_short_wrap.wav')
+        FileUtils.rm_f(result_file_path)
+        processor.on_message(msg)
+        File.exists?(result_file_path).must_equal true
+        wave = NuWav::WaveFile.parse(result_file_path)
 
-      wave.chunks[:cart].title.must_equal "REMIX Episode 1"
-      wave.chunks[:cart].artist.must_equal "PRX REMIX"
-      wave.chunks[:cart].cut_id.must_equal "12345"
-      wave.chunks[:cart].start_date.must_equal "2010/06/19"
-      wave.chunks[:cart].start_time.must_equal "00:00:00"
-      wave.chunks[:cart].end_date.must_equal "2010/06/25"
-      wave.chunks[:cart].end_time.must_equal "00:00:00"
-      wave.chunks[:cart].producer_app_id.must_equal "PRX"
+        wave.chunks[:cart].title.must_equal "REMIX Episode 1"
+        wave.chunks[:cart].artist.must_equal "PRX REMIX"
+        wave.chunks[:cart].cut_id.must_equal "12345"
+        wave.chunks[:cart].start_date.must_equal "2010/06/19"
+        wave.chunks[:cart].start_time.must_equal "00:00:00"
+        wave.chunks[:cart].end_date.must_equal "2010/06/25"
+        wave.chunks[:cart].end_time.must_equal "00:00:00"
+        wave.chunks[:cart].producer_app_id.must_equal "PRX"
+      end
     end
   end
 
@@ -182,20 +232,39 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it "should generate waveform" do
-      processor.on_message(msg)
-      processor.destination.wont_be_nil
-      processor.destination_format.must_equal 'json'
-      processor.result_details[:info].keys.sort.must_equal [:data_count]
-      processor.result_details[:message].wont_be :blank?
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:encode_wav_pcm_from_mp2, ["0\n", ""], [String, String])
+      end
+
+      processor.stub(:waveformjson, { foo: 'bar'} ) do
+        processor.on_message(msg)
+        processor.destination.wont_be_nil
+        processor.destination_format.must_equal 'json'
+        processor.result_details[:info].keys.sort.must_equal [:data_count]
+        processor.result_details[:message].wont_be :blank?
+      end
     end
 
     it "should use with per second option" do
-      msg[:task][:options] = { width_per_second: 500 }
-      processor.on_message(msg)
-      processor.destination.wont_be_nil
-      processor.destination_format.must_equal 'json'
-      processor.result_details[:info].keys.sort.must_equal [:data_count]
-      processor.result_details[:info][:data_count].must_be :>=, 2800
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:encode_wav_pcm_from_mp2, ["0\n", ""], [String, String])
+        audio_monster.expect(:audio_file_duration, 5.616, [String])
+      end
+
+      result = {}
+      2800.times { |i| result[i] = i }
+      processor.stub(:waveformjson, result ) do
+        msg[:task][:options] = { width_per_second: 500 }
+        processor.on_message(msg)
+        processor.destination.wont_be_nil
+        processor.destination_format.must_equal 'json'
+        processor.result_details[:info].keys.sort.must_equal [:data_count]
+        processor.result_details[:info][:data_count].must_be :>=, 2800
+      end
     end
   end
 
@@ -213,14 +282,21 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it "should get json results" do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String, false])
+        audio_monster.expect(:encode_wav_pcm_from_mp2, ["0\n", ""], [String, String])
+      end
 
-      stub_request(:post, "https://www.google.com/speech-api/v2/recognize?client=chrome&key=AIzaSyBOti4mM-6x9WDnZIjIeyEU21OpBXqWBgw&lang=en-us&output=json").
-        with(:headers => { 'Content-Type'=>'audio/x-flac; rate=8000', 'Host'=>'www.google.com:443' } ).
-        to_return(:status => 200, :body => "", :headers => {})
+      # stub_request(:post, "https://www.google.com/speech-api/v2/recognize?client=chrome&key=AIzaSyBOti4mM-6x9WDnZIjIeyEU21OpBXqWBgw&lang=en-us&output=json").
+      #   with(:headers => { 'Content-Type'=>'audio/x-flac; rate=8000', 'Host'=>'www.google.com:443' } ).
+      #   to_return(:status => 200, :body => "", :headers => {})
 
-      processor.on_message(msg)
-      processor.result_details[:info].keys.sort.must_equal [:average_confidence, :character_count, :word_count]
-      processor.result_details[:message].wont_be_nil
+      processor.stub(:google_transcribe, [ { text: 'the quick brown fox jumped over the lazy dog', confidence: 1.0 } ]) do
+        processor.on_message(msg)
+        processor.result_details[:info].keys.sort.must_equal [:average_confidence, :character_count, :word_count]
+        processor.result_details[:message].wont_be_nil
+      end
     end
   end
 
@@ -237,6 +313,17 @@ class AudioProcessorTest < ActiveSupport::TestCase
           result: 'file:///test.fixer.org/test_short_test.mp3'
         }
       }.with_indifferent_access
+    }
+
+    before {
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:encode_wav_pcm_from_mp2, ["0\n", ""], [String, String])
+        audio_monster.expect(:encode_mp3_from_wav, true, [String, String, Hash])
+        audio_monster.expect(:info_for_mp3, {:size=>90696, :content_type=>"audio/mpeg", :channel_mode=>"JStereo", :bit_rate=>128, :length=>5, :sample_rate=>44100, :version=>1, :layer=>3}, [String])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+      end
     }
 
     it "should set destination format" do
@@ -292,13 +379,20 @@ class AudioProcessorTest < ActiveSupport::TestCase
       }.with_indifferent_access
     }
 
+    before {
+      if travis?
+        audio_monster.expect(:info_for_mp2, {:size=>179712, :content_type=>"audio/mpeg", :channel_mode=>"Stereo", :bit_rate=>256, :length=>5, :sample_rate=>48000, :version=>1, :layer=>2}, [String])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+      end
+    }
+
     it "should set destination format" do
       processor.on_message(msg)
       processor.destination.wont_be_nil
       processor.destination_format.must_equal 'mp2'
     end
 
-    it "should create a wave wrapped mp2" do
+    it "should copy the file" do
       result_file_path = out_file('fixer/test.fixer.org/test_short_copy.mp2')
       FileUtils.rm_f(result_file_path)
       processor.on_message(msg)
@@ -333,6 +427,11 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it "should validate" do
+      if travis?
+        audio_monster.expect(:info_for_mp2, {:size=>179712, :content_type=>"audio/mpeg", :channel_mode=>"Stereo", :bit_rate=>256, :length=>5, :sample_rate=>48000, :version=>1, :layer=>2}, [String])
+        audio_monster.expect(:validate_mp2, [{}, {:size=>179712, :content_type=>"audio/mpeg", :channel_mode=>"Stereo", :bit_rate=>256, :length=>5, :sample_rate=>48000, :version=>1, :layer=>2}], [String, Hash])
+      end
+
       processor.on_message(msg)
       # {:status=>:complete, :info=>{:errors=>{}, :analysis=>{:size=>179712, :content_type=>"audio/mpeg", :channel_mode=>"Stereo", :bit_rate=>256, :length=>5, :sample_rate=>48000, :version=>1, :layer=>2}}, :message=>"Validate audio complete."}
       processor.result_details[:info][:errors].keys.length.must_equal 0
@@ -355,6 +454,12 @@ class AudioProcessorTest < ActiveSupport::TestCase
     }
 
     it 'should cut' do
+      if travis?
+        audio_monster.expect(:create_temp_file, Tempfile.new('test'), [String])
+        audio_monster.expect(:info_for_wav, {:size=>4277159, :content_type=>"audio/vnd.wave", :channel_mode=>"Mono", :bit_rate=>705, :length=>48, :sample_rate=>44100}, [String])
+        audio_monster.expect(:loudness_info, {:integrated_loudness=>{:i=>-18.5, :threshold=>-28.6}, :loudness_range=>{:lra=>7.1, :threshold=>-38.7, :lra_low=>-23.6, :lra_high=>-16.5}, :true_peak=>{:peak=>-2.1}}, [String])
+        audio_monster.expect(:cut_wav, ["0", "sox WARN sox: Option `-s' is deprecated, use `-e signed-integer' instead.", "sox WARN sox: Option `-s' is deprecated, use `-e signed-integer' instead."], [String, String, Fixnum, Fixnum])
+      end
       processor.on_message(msg)
     end
   end
